refactor: DTO의 의존성 흐름 변경
