+++ src/main/java/subway/business/service/dto/StationAddToLineRequest.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.business.service.dto;
 
 public class StationAddToLineRequest {
     private final String station;
