+++ src/main/java/subway/ui/dto/StationDeleteRequest.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.ui.dto;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
